refacto & delete point from right click and from timeline trash button on drawPath
