+++ src/views/content/path/component/atom/TimeLineCircle.css
@@ -1,5 +1,14 @@
 .path-timeline-circle {
-  @apply absolute top-[-15px] left-1/2 w-[35px] h-[35px] rounded-2xl;
+  @apply flex items-center justify-center absolute top-[-15px] left-1/2 w-[35px] h-[35px] rounded-full;
   background-color: red;
   transform: translateX(-50%);
+}
+
+.timeline-delete-point {
+  @apply opacity-0 hover:opacity-100;
+  height: calc(100% - 8px);
+  width: calc(100% - 8px);
+}
+.timeline-delete-point > svg {
+  @apply rounded-full bg-white fill-red-base h-full w-full p-[6px];
 }
